Split project across files, add .vscode to gitignore
