+++ schedule-bot/dateutils.py
@@ -1,0 +1,14 @@
+import datetime
+from time import mktime
+
+
+def struct_to_datetime(value):
+    return datetime.datetime.fromtimestamp(mktime(value))
+
+
+def datetimeformat(value, format="%Y-%m-%d"):
+    return value.strftime(format)
+
+
+def structformat(value, format="%Y-%m-%d"):
+    return datetimeformat(struct_to_datetime(value), format)
